squash some easy warnings

+++ src/Three20Style/Sources/UIColorAdditions.m
@@ -117,7 +117,7 @@ TT_FIX_CATEGORY_BUG(UIColorAdditions)
 
 ///////////////////////////////////////////////////////////////////////////////////////////////////
 + (UIColor*)colorWithHue:(CGFloat)h saturation:(CGFloat)s value:(CGFloat)v alpha:(CGFloat)a {
-  CGFloat r, g, b;
+  float r, g, b;
   HSVtoRGB(&r, &g, &b, h, s, v);
   return [UIColor colorWithRed:r green:g blue:b alpha:a];
 }
@@ -126,12 +126,12 @@ TT_FIX_CATEGORY_BUG(UIColorAdditions)
 ///////////////////////////////////////////////////////////////////////////////////////////////////
 - (UIColor*)multiplyHue:(CGFloat)hd saturation:(CGFloat)sd value:(CGFloat)vd {
   const CGFloat* rgba = CGColorGetComponents(self.CGColor);
-  CGFloat r = rgba[0];
-  CGFloat g = rgba[1];
-  CGFloat b = rgba[2];
-  CGFloat a = rgba[3];
+  float r = rgba[0];
+  float g = rgba[1];
+  float b = rgba[2];
+  float a = rgba[3];
 
-  CGFloat h, s, v;
+  float h, s, v;
   RGBtoHSV(r, g, b, &h, &s, &v);
 
   h *= hd;
@@ -158,12 +158,12 @@ TT_FIX_CATEGORY_BUG(UIColorAdditions)
 ///////////////////////////////////////////////////////////////////////////////////////////////////
 - (UIColor*)addHue:(CGFloat)hd saturation:(CGFloat)sd value:(CGFloat)vd {
   const CGFloat* rgba = CGColorGetComponents(self.CGColor);
-  CGFloat r = rgba[0];
-  CGFloat g = rgba[1];
-  CGFloat b = rgba[2];
-  CGFloat a = rgba[3];
+  float r = rgba[0];
+  float g = rgba[1];
+  float b = rgba[2];
+  float a = rgba[3];
 
-  CGFloat h, s, v;
+  float h, s, v;
   RGBtoHSV(r, g, b, &h, &s, &v);
 
   h += hd;
@@ -191,7 +191,7 @@ TT_FIX_CATEGORY_BUG(UIColorAdditions)
 ///////////////////////////////////////////////////////////////////////////////////////////////////
 - (CGFloat)hue {
   const CGFloat* rgba = CGColorGetComponents(self.CGColor);
-  CGFloat h, s, v;
+  float h, s, v;
   RGBtoHSV(rgba[0], rgba[1], rgba[2], &h, &s, &v);
   return h;
 }
@@ -200,7 +200,7 @@ TT_FIX_CATEGORY_BUG(UIColorAdditions)
 ///////////////////////////////////////////////////////////////////////////////////////////////////
 - (CGFloat)saturation {
   const CGFloat* rgba = CGColorGetComponents(self.CGColor);
-  CGFloat h, s, v;
+  float h, s, v;
   RGBtoHSV(rgba[0], rgba[1], rgba[2], &h, &s, &v);
   return s;
 }
@@ -209,7 +209,7 @@ TT_FIX_CATEGORY_BUG(UIColorAdditions)
 ///////////////////////////////////////////////////////////////////////////////////////////////////
 - (CGFloat)value {
   const CGFloat* rgba = CGColorGetComponents(self.CGColor);
-  CGFloat h, s, v;
+  float h, s, v;
   RGBtoHSV(rgba[0], rgba[1], rgba[2], &h, &s, &v);
   return v;
 }
